pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
marshmallow = "^3.20.2"
flask-smorest = "^0.43.0"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.5"
pytest-mock = "^3.12.0"
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
mysql-connector-python = "^8.3.0"
flask-login = "^0.6.3"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
